Fix: Ensure $item is an Asset before getting original aspect ratio
Fixes a bug where a String (like a path) passed into the `wildcard` function would result in a type error. 

This was caused by the `getOriginalAspectRatio` call in `getFinalDimensions` because `$item` is never converted to an `AssetsAsset` (via `getAsset()`).

Another way to fix this would be to ensure `$item` is an `AssetsAsset` before this point, but this fix is more consistent with the code in `generateCloudinaryUrl`.